refactor: replace old storage system and related cleanup
