+++ src/routes/storage/index.ts
@@ -1,13 +1,35 @@
-import { Router } from 'express'
+import { Router, NextFunction, Request, Response } from 'express'
 
-import getFile from './file'
-import getFileMeta from './meta'
-import deleteFile from './delete'
-import uploadFile from './upload'
-import revokeFile from './revoke'
-export default Router()
-  .delete('/file/*', deleteFile)
-  .get('/file-meta/*', getFileMeta)
-  .get('/file/*', getFile)
-  .post('/revoke/*', revokeFile)
-  .post('/upload', uploadFile)
+import { STORAGE_RULES } from './utils'
+import { uploadFile } from './upload'
+import { getFile } from './get'
+import { deleteFile } from './delete'
+const router = Router()
+
+const createSecureMiddleware = (fn: Function, rules: (string | undefined)[]) => (
+  req: Request,
+  res: Response,
+  next: NextFunction
+): void => fn(req, res, next, rules).catch(next)
+
+for (const path in STORAGE_RULES.paths) {
+  const rule = STORAGE_RULES.paths[path]
+  if (rule.create || rule.update || rule.write) {
+    router.post(path, createSecureMiddleware(uploadFile, [rule.create, rule.update, rule.write]))
+  }
+  if (rule.get || rule.read) {
+    router.get(path, createSecureMiddleware(getFile, [rule.get, rule.read]))
+  }
+  // TODO list
+  // if (rule.list || rule.read) {
+  //   router.get(`${path}/`, () => {
+  //     console.log('TODO')
+  //   })
+  // }
+  if (rule.delete || rule.write) {
+    router.delete(path, createSecureMiddleware(deleteFile, [rule.delete, rule.write]))
+  }
+  // TODO metadata
+}
+
+export default router
