+++ src/routes/storage/upload.ts
@@ -1,62 +1,51 @@
-import { Request, Response } from 'express'
+import { Request, Response, NextFunction } from 'express'
 
 import Boom from '@hapi/boom'
 import { S3_BUCKET } from '@shared/config'
 import { UploadedFile } from 'express-fileupload'
-import { asyncWrapper } from '@shared/helpers'
 import { s3 } from '@shared/s3'
-import { storagePermission } from '@custom/storage-rules'
-import { v4 as uuidv4 } from 'uuid'
-import { verify } from '@shared/jwt'
+import { hasPermission, createContext, getKey } from './utils'
 
-async function uploadFile(req: Request, res: Response): Promise<unknown> {
+export const uploadFile = async (
+  req: Request,
+  res: Response,
+  _next: NextFunction,
+  rules: (string | undefined)[]
+): Promise<unknown> => {
+  const key = getKey(req)
+
   if (!req.files?.file) {
     throw Boom.notFound()
   }
 
-  // get file being uploaded
-  const uploaded_file = req.files.file as UploadedFile
+  const resource = req.files.file as UploadedFile
+  const context = createContext(req, resource)
 
-  // get path to upload file
-  const key = req.headers['x-key'] as string
-
-  // check storageRules if request is ok
-  const jwt_token = verify(req.headers.authorization)
-  const claims = jwt_token['https://hasura.io/jwt/claims']
-
-  if (!storagePermission(key, 'write', claims)) {
+  if (!hasPermission(rules, context)) {
     throw Boom.forbidden()
   }
-
-  // generate access token for file
-  const token = uuidv4()
-
-  // upload file
   const upload_params = {
     Bucket: S3_BUCKET as string,
     Key: key,
-    Body: uploaded_file.data,
-    ContentType: uploaded_file.mimetype,
+    Body: resource.data,
+    ContentType: resource.mimetype,
     Metadata: {
-      token,
-      filename: uploaded_file.name
+      filename: resource.name
     }
   }
 
   try {
     await s3.upload(upload_params).promise()
   } catch (err) {
+    console.log('upload pb')
+    console.log(err)
     throw Boom.badImplementation()
   }
 
   // return info about the uploaded file
-  const { name, mimetype } = uploaded_file
   return res.status(200).send({
-    key,
-    filename: name,
-    mimetype,
-    token
+    path: key,
+    filename: resource.name,
+    mimetype: resource.mimetype
   })
 }
-
-export default asyncWrapper(uploadFile)
